# Conflicts:
#	.gitignore
#	IGraphics/IGraphics.cpp
#	IGraphics/IGraphicsEditorDelegate.h
#	IGraphics/Platforms/IGraphicsMac.mm
#	IPlug/APP/IPlugAPP.cpp
#	IPlug/APP/IPlugAPP_dialog.cpp
#	IPlug/IPlugAPIBase.h
#	IPlug/IPlugEditorDelegate.h
#	IPlug/IPlugPaths.cpp
#	IPlug/VST2/IPlugVST2.cpp

+++ IGraphics/Drawing/IGraphicsSkia.h
@@ -116,7 +116,7 @@ public:
   int AlphaChannel() const override { return 3; }
   bool FlippedBitmap() const override { return false; }
 
-  APIBitmap* CreateAPIBitmap(int width, int height, int scale, double drawScale, bool cacheable = false) override;
+  APIBitmap* CreateAPIBitmap(int width, int height, float scale, double drawScale, bool cacheable = false) override;
 
   void GetLayerBitmapData(const ILayerPtr& layer, RawBitmapData& data) override;
   void ApplyShadowMask(ILayerPtr& layer, RawBitmapData& mask, const IShadow& shadow) override;
@@ -132,8 +132,7 @@ protected:
 
   APIBitmap* LoadAPIBitmap(const char* fileNameOrResID, int scale, EResourceLocation location, const char* ext) override;
   APIBitmap* LoadAPIBitmap(const char* name, const void* pData, int dataSize, int scale) override;
-private:
-    
+private:  
   void PrepareAndMeasureText(const IText& text, const char* str, IRECT& r, double& x, double & y, SkFont& font) const;
 
   void PathTransformSetMatrix(const IMatrix& m) override;
@@ -163,7 +162,7 @@ private:
   void* mMTLDrawable;
   void* mMTLLayer;
 #endif
-  
+
   static StaticStorage<Font> sFontCache;
 };
